evaluations/pyproject.toml
[project]
name = "azure-ai-evaluation-tryout"
version = "0.1.0"
description = "A project for evaluating AI models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "azure-ai-evaluation>=1.7.0",
    "azure-ai-projects>=1.0.0b11",
    "azure-monitor-opentelemetry>=1.6.9",
    "opentelemetry-sdk>=1.31.1",
    "python-dotenv>=1.1.0",
]
